Add the plugins folder
The aim was to tidy my plugins file and put all the plugin information
in the plugin specific file which is much tidier.

I had to remove the .gitignore and also install a nerdfont to make the
icons work correctly with neotree.

Overall much happier with this setup and more to come.